FeatureToolbar.tsx をFeatureManager.tsxに統合
